backend/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
axblack==20220330
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==3.2.3
django-filter==23.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
idna==3.4
install==1.3.5
mypy-extensions==1.0.0
oauthlib==3.2.2
pathspec==0.11.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
toml==0.10.2
typed-ast==1.5.5
typing_extensions==4.8.0
urllib3==2.1.0
